moved tests to appropriate test file
